pyproject.toml
[project]
name = "spacegom-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.30",
    "botocore>=1.42.30",
    "fastapi>=0.128.0",
    "python-dotenv>=1.0.0",
    "flake8>=7.3.0",
    "jinja2>=3.1.6",
    "mypy>=1.19.1",
    "numpy>=2.0.0,<2.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.0.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[project.scripts]
dev = "app.run:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pylint>=4.0.4",
]
